check name change
