+++ deepchecks/tabular/checks/performance/__init__.py
@@ -18,11 +18,11 @@ Module contains checks of model performance metrics.
 import warnings
 
 from ..model_evaluation import (CalibrationScore, ConfusionMatrixReport, ModelErrorAnalysis,
-                                MultiModelPerformanceReport, PerformanceReport, RegressionErrorDistribution,
+                                MultiModelPerformanceReport, TrainTestPerformance, RegressionErrorDistribution,
                                 RegressionSystematicError, RocReport, SegmentPerformance, SimpleModelComparison)
 
 __all__ = [
-    'PerformanceReport',
+    'TrainTestPerformance',
     'ConfusionMatrixReport',
     'RocReport',
     'SimpleModelComparison',
