+++ deepchecks/tabular/suites/default_suites.py
@@ -24,7 +24,7 @@ from deepchecks.tabular.checks import (BoostingOverfit, CalibrationScore, Catego
                                        FeatureLabelCorrelation, FeatureLabelCorrelationChange,
                                        IdentifierLabelCorrelation, IndexTrainTestLeakage, IsSingleValue, MixedDataTypes,
                                        MixedNulls, ModelInferenceTime, NewLabelTrainTest, OutlierSampleDetection,
-                                       PerformanceReport, RegressionErrorDistribution, RegressionSystematicError,
+                                       TrainTestPerformance, RegressionErrorDistribution, RegressionSystematicError,
                                        RocReport, SimpleModelComparison, SpecialCharacters, StringLengthOutOfBounds,
                                        StringMismatch, StringMismatchComparison, TrainTestFeatureDrift,
                                        TrainTestLabelDrift, TrainTestPredictionDrift, TrainTestSamplesMix,
@@ -344,7 +344,7 @@ def model_evaluation(alternative_scorers: Dict[str, Callable] = None,
 
     return Suite(
         'Model Evaluation Suite',
-        PerformanceReport(**kwargs).add_condition_train_test_relative_degradation_less_than(),
+        TrainTestPerformance(**kwargs).add_condition_train_test_relative_degradation_less_than(),
         RocReport(**kwargs).add_condition_auc_greater_than(),
         ConfusionMatrixReport(**kwargs),
         TrainTestPredictionDrift(**kwargs).add_condition_drift_score_less_than(),
