+++ tests/tabular/checks/model_evaluation/train_test_performance_test.py
@@ -21,7 +21,7 @@ from sklearn.model_selection import train_test_split
 from deepchecks.core import ConditionResult
 from deepchecks.core.errors import (DatasetValidationError, DeepchecksNotSupportedError, DeepchecksValueError,
                                     ModelValidationError)
-from deepchecks.tabular.checks.model_evaluation import PerformanceReport
+from deepchecks.tabular.checks.model_evaluation import TrainTestPerformance
 from deepchecks.tabular.dataset import Dataset
 from deepchecks.tabular.metric_utils.scorers import DEFAULT_REGRESSION_SCORERS, MULTICLASS_SCORERS_NON_AVERAGE
 from tests.base.utils import equal_condition_result
@@ -31,7 +31,7 @@ def test_dataset_wrong_input():
     bad_dataset = 'wrong_input'
     # Act & Assert
     assert_that(
-        calling(PerformanceReport().run).with_args(bad_dataset, None, None),
+        calling(TrainTestPerformance().run).with_args(bad_dataset, None, None),
         raises(DeepchecksValueError, 'non-empty instance of Dataset or DataFrame was expected, instead got str')
     )
 
@@ -40,7 +40,7 @@ def test_model_wrong_input(iris_labeled_dataset):
     bad_model = 'wrong_input'
     # Act & Assert
     assert_that(
-        calling(PerformanceReport().run).with_args(iris_labeled_dataset, iris_labeled_dataset, bad_model),
+        calling(TrainTestPerformance().run).with_args(iris_labeled_dataset, iris_labeled_dataset, bad_model),
         raises(
             ModelValidationError,
             r'Model supplied does not meets the minimal interface requirements. Read more about .*')
@@ -50,7 +50,7 @@ def test_model_wrong_input(iris_labeled_dataset):
 def test_dataset_no_label(iris_dataset_no_label, iris_adaboost):
     # Assert
     assert_that(
-        calling(PerformanceReport().run).with_args(iris_dataset_no_label, iris_dataset_no_label, iris_adaboost),
+        calling(TrainTestPerformance().run).with_args(iris_dataset_no_label, iris_dataset_no_label, iris_adaboost),
         raises(DeepchecksNotSupportedError,
                'Dataset does not contain a label column')
     )
@@ -60,7 +60,7 @@ def test_dataset_no_shared_label(iris_labeled_dataset):
     # Assert
     iris_dataset_2 = Dataset(iris_labeled_dataset.data, label='sepal length (cm)')
     assert_that(
-        calling(PerformanceReport().run).with_args(iris_labeled_dataset, iris_dataset_2, None),
+        calling(TrainTestPerformance().run).with_args(iris_labeled_dataset, iris_dataset_2, None),
         raises(DatasetValidationError, 'train and test requires to have and to share the same label')
     )
 
@@ -78,7 +78,7 @@ def assert_classification_result(result, dataset: Dataset):
 def test_classification(iris_split_dataset_and_model):
     # Arrange
     train, test, model = iris_split_dataset_and_model
-    check = PerformanceReport()
+    check = TrainTestPerformance()
     # Act
     result = check.run(train, test, model)
     # Assert
@@ -89,7 +89,7 @@ def test_classification(iris_split_dataset_and_model):
 def test_classification_without_display(iris_split_dataset_and_model):
     # Arrange
     train, test, model = iris_split_dataset_and_model
-    check = PerformanceReport()
+    check = TrainTestPerformance()
     # Act
     result = check.run(train, test, model, with_display=False)
     # Assert
@@ -104,7 +104,7 @@ def test_classification_binary(iris_dataset_single_class_labeled):
     test_ds = iris_dataset_single_class_labeled.copy(test)
     clf = RandomForestClassifier(random_state=0)
     clf.fit(train_ds.data[train_ds.features], train_ds.data[train_ds.label_name])
-    check = PerformanceReport()
+    check = TrainTestPerformance()
 
     # Act
     result = check.run(train_ds, test_ds, clf).value
@@ -114,7 +114,7 @@ def test_classification_binary(iris_dataset_single_class_labeled):
 
 def test_classification_string_labels(iris_labeled_dataset):
     # Arrange
-    check = PerformanceReport()
+    check = TrainTestPerformance()
     replace_dict = {iris_labeled_dataset.label_name: {0: 'b', 1: 'e', 2: 'a'}}
     iris_labeled_dataset = Dataset(iris_labeled_dataset.data.replace(replace_dict),
                                    label=iris_labeled_dataset.label_name)
@@ -130,7 +130,7 @@ def test_classification_string_labels(iris_labeled_dataset):
 
 def test_classification_nan_labels(iris_labeled_dataset, iris_adaboost):
     # Arrange
-    check = PerformanceReport()
+    check = TrainTestPerformance()
     data_with_nan = iris_labeled_dataset.data.copy()
     data_with_nan[iris_labeled_dataset.label_name].iloc[0] = float('nan')
     iris_labeled_dataset = Dataset(data_with_nan,
@@ -144,7 +144,7 @@ def test_classification_nan_labels(iris_labeled_dataset, iris_adaboost):
 def test_regression(diabetes_split_dataset_and_model):
     # Arrange
     train, test, model = diabetes_split_dataset_and_model
-    check = PerformanceReport()
+    check = TrainTestPerformance()
     # Act
     result = check.run(train, test, model).value
     # Assert
@@ -158,7 +158,7 @@ def test_regression(diabetes_split_dataset_and_model):
 def test_regression_reduced(diabetes_split_dataset_and_model):
     # Arrange
     train, test, model = diabetes_split_dataset_and_model
-    check = PerformanceReport()
+    check = TrainTestPerformance()
     # Act
     result = check.run(train, test, model).reduce_output()
     # Assert
@@ -170,7 +170,7 @@ def test_regression_reduced(diabetes_split_dataset_and_model):
 def test_classification_reduced(iris_split_dataset_and_model):
     # Arrange
     train, test, model = iris_split_dataset_and_model
-    check = PerformanceReport()
+    check = TrainTestPerformance()
     # Act
     result = check.run(train, test, model).reduce_output()
     # Assert
@@ -182,7 +182,7 @@ def test_classification_reduced(iris_split_dataset_and_model):
 def test_classification_reduced_param(iris_split_dataset_and_model):
     # Arrange
     train, test, model = iris_split_dataset_and_model
-    check = PerformanceReport(reduce=np.min)
+    check = TrainTestPerformance(reduce=np.min)
     # Act
     result = check.run(train, test, model).reduce_output()
     # Assert
@@ -194,7 +194,7 @@ def test_classification_reduced_param(iris_split_dataset_and_model):
 def test_condition_min_score_not_passed(iris_split_dataset_and_model):
     # Arrange
     train, test, model = iris_split_dataset_and_model
-    check = PerformanceReport().add_condition_test_performance_greater_than(1)
+    check = TrainTestPerformance().add_condition_test_performance_greater_than(1)
     # Act
     result: List[ConditionResult] = check.conditions_decision(check.run(train, test, model))
     # Assert
@@ -209,7 +209,7 @@ def test_condition_min_score_not_passed(iris_split_dataset_and_model):
 def test_condition_min_score_passed(iris_split_dataset_and_model):
     # Arrange
     train, test, model = iris_split_dataset_and_model
-    check = PerformanceReport().add_condition_test_performance_greater_than(0.5)
+    check = TrainTestPerformance().add_condition_test_performance_greater_than(0.5)
     # Act
     result: List[ConditionResult] = check.conditions_decision(check.run(train, test, model))
     # Assert
@@ -223,7 +223,7 @@ def test_condition_min_score_passed(iris_split_dataset_and_model):
 def test_condition_degradation_ratio_less_than_not_passed(iris_split_dataset_and_model):
     # Arrange
     train, test, model = iris_split_dataset_and_model
-    check = PerformanceReport().add_condition_train_test_relative_degradation_less_than(0)
+    check = TrainTestPerformance().add_condition_train_test_relative_degradation_less_than(0)
     # Act
     result: List[ConditionResult] = check.conditions_decision(check.run(train, test, model))
     # Assert
@@ -237,7 +237,7 @@ def test_condition_degradation_ratio_less_than_not_passed(iris_split_dataset_and
 def test_condition_degradation_ratio_less_than_passed(iris_split_dataset_and_model):
     # Arrange
     train, test, model = iris_split_dataset_and_model
-    check = PerformanceReport().add_condition_train_test_relative_degradation_less_than(1)
+    check = TrainTestPerformance().add_condition_train_test_relative_degradation_less_than(1)
     # Act
     result: List[ConditionResult] = check.conditions_decision(check.run(train, test, model))
     # Assert
@@ -251,7 +251,7 @@ def test_condition_degradation_ratio_less_than_passed(iris_split_dataset_and_mod
 def test_condition_class_performance_imbalance_ratio_less_than_not_passed(iris_split_dataset_and_model):
     # ArrangeF
     train, test, model = iris_split_dataset_and_model
-    check = PerformanceReport().add_condition_class_performance_imbalance_ratio_less_than(0)
+    check = TrainTestPerformance().add_condition_class_performance_imbalance_ratio_less_than(0)
     # Act
     result: List[ConditionResult] = check.conditions_decision(check.run(train, test, model))
     # Assert
@@ -266,7 +266,7 @@ def test_condition_class_performance_imbalance_ratio_less_than_not_passed(iris_s
 def test_condition_class_performance_imbalance_ratio_less_than_passed(iris_split_dataset_and_model):
     # Arrange
     train, test, model = iris_split_dataset_and_model
-    check = PerformanceReport().add_condition_class_performance_imbalance_ratio_less_than(1)
+    check = TrainTestPerformance().add_condition_class_performance_imbalance_ratio_less_than(1)
     # Act
     result: List[ConditionResult] = check.conditions_decision(check.run(train, test, model))
     # Assert
@@ -281,7 +281,7 @@ def test_condition_class_performance_imbalance_ratio_less_than_passed(iris_split
 def test_classification_alt_scores_list(iris_split_dataset_and_model):
     # Arrange
     train, test, model = iris_split_dataset_and_model
-    check = PerformanceReport(scorers=['recall_per_class',
+    check = TrainTestPerformance(scorers=['recall_per_class',
                               'f1_per_class', make_scorer(jaccard_score, average=None)])
     # Act
     result = check.run(train, test, model).reduce_output()
@@ -294,7 +294,7 @@ def test_classification_alt_scores_list(iris_split_dataset_and_model):
 def test_classification_deepchecks_scorers(iris_split_dataset_and_model):
     # Arrange
     train, test, model = iris_split_dataset_and_model
-    check = PerformanceReport(scorers=['fpr_per_class', 'fnr_per_class', 'specificity_per_class'])
+    check = TrainTestPerformance(scorers=['fpr_per_class', 'fnr_per_class', 'specificity_per_class'])
     # Act
     result = check.run(train, test, model).reduce_output()
     # Assert
@@ -306,7 +306,7 @@ def test_classification_deepchecks_scorers(iris_split_dataset_and_model):
 def test_regression_alt_scores_list(diabetes_split_dataset_and_model):
     # Arrange
     train, test, model = diabetes_split_dataset_and_model
-    check = PerformanceReport(scorers=['max_error', 'r2', 'neg_mean_absolute_error'])
+    check = TrainTestPerformance(scorers=['max_error', 'r2', 'neg_mean_absolute_error'])
     # Act
     result = check.run(train, test, model).reduce_output()
     # Assert
@@ -318,7 +318,7 @@ def test_regression_alt_scores_list(diabetes_split_dataset_and_model):
 def test_classification_alt_scores_per_class_and_macro(iris_split_dataset_and_model):
     # Arrange
     train, test, model = iris_split_dataset_and_model
-    check = PerformanceReport(scorers=['recall_per_class', 'f1_per_class', 'f1_macro', 'recall_micro'])
+    check = TrainTestPerformance(scorers=['recall_per_class', 'f1_per_class', 'f1_macro', 'recall_micro'])
     # Act
     result = check.run(train, test, model).reduce_output()
     # Assert
